datacube: reorganize application engine
